fixes for cases of total mass missing: support payload-only input (no mass_total) in evolver pipeline
- make_population: fix crash when mass_total absent — replaced invalid
  cell indexing (derived{2}) with correct struct path lookup via
  derived.unknown.mass.mass_total / derived.known.mass.mass_total

- make_population: don't apply mass_payload / power_payload as hard
  evolver floors when mass_total is absent; in payload-only mode these
  fields drive the mass estimate and must not additionally constrain
  mutations (caused all 18 lineages to stall)

- input_processing: suppress misleading 'Total mass limit exceeded'
  warning in system_with_unknown_totals path — added warn_if_exceeded
  flag (default true) to mass_validate; passes false when total was
  derived, not user-given

- evolver: add error when all lineages converge as stall with no
  successful mutation, with diagnostic message listing likely causes

- evolve_population: cast count_for_convergence to double — logical
  type from && was not handled by val2str in struct2xml, producing
  spurious XML serialisation warnings

+++ Code/Evolver/evolve_population.m
@@ -190,7 +190,7 @@ function [generation_new, convergence] = evolve_population(input, db_data, confi
       elseif abs(population(i,j).thrust - parent_individual.thrust) > max(1, abs(parent_individual.thrust))*1e-12
         mutation_changed = true;
       end
-      population(i,j).count_for_convergence = mutation_valid && mutation_changed;
+      population(i,j).count_for_convergence = double(mutation_valid && mutation_changed);
       population(i,j).convergence_mode = 'active';
 
       %refresh the number of the last sucessful lineage member here
